Control the cap for plugins and the core separately.
For example, UC for mainline Jenkins (1.532,1.545] should only show plugins that are compatible up to 1.545,
but it should still advertise the latest core release, say 1.590.

This is different from UC for LTS Jenkins 1.545.x, which should only show plugins that are compatible up to 1.545.x,
 and it should only advertise core up to 1.545.x

To control this separately, we need two switches.