Exclude plugins without a valid wiki page from the Update Centre JSON files.
The criteria for inclusion are:
- Plugins must specify a Jenkins wiki URL in their pom.xml, via the <url> tag
- The wiki page must exist on the Jenkins wiki
- The wiki page must be a child of the "Plugins" page

Also accepted are:
- Wiki short URLs (e.g. /x/GYCGAQ)
- Wiki URLs with a numeric ID (e.g. /pages/viewpage.action?pageId=60915753)
- URLs which have legacy URL prefixes (e.g. non-HTTPS, wiki.hudson-ci.org)
- URLs ending with a trailing slash

Plugins not meeting these criteria will not be included in the Update Centre
JSON file, nor in the release history JSON file.

As before, plugins may also be excluded via the "artifact-ignores.properties"
file, or if a plugin does specify a valid wiki page, but that wiki page has the
"plugin-deprecated" label.

While building the plugin list, more informational output is now printed to the
console regarding why plugins are being excluded, or whether their wiki URL was
rewritten to the canonical wiki URL format.

A summary of the total number of included, excluded and deprecated plugins is
also shown at the end of building the Update Centre plugin list.

A number of test cases were added to test the basic URL resolution process, as
well as some of the rarer types of URL found in currently-released plugins.