When building the old/stable UC files, keep plugins without wiki pages for now.
Even if plugin developers update their plugins to include a wiki URL, and make a
new release, this will not apply to older plugin releases.  Because the UC for
LTS releases may only include older plugin releases, and not the latest version
of any given plugin, we should not exclude those plugins because they don't have
a wiki page.

This is a workaround which should ideally be removed once we no longer generate
UC files for LTS 1.596, which was the last LTS in which we tolerated plugins
without a valid wiki URL.