Don't exclude plugins whose wiki pages temporarily could not be fetched.
Also be smarter about handling invalid wiki page cache files; until now, we
were not re-attempting to fetch a wiki page if there was *any* sort of cache
file in existence, even if it was empty or contained a `null` object reference.

Added more logging to cover these and related situations.