Adds an option to just verify that a plugin would be included.
Given the argument "-verify-plugin someString", no files will actually
be written, core won't be processed, and only plugins containing
'someString' will be processed. This can be used to verify whether a
plugin release has a valid wiki page and meets any other criteria we
may add in the future.

Usage:

- mvn clean install
- java -jar target/update-center2-*-bin*/update-center2-*.jar \
    -id testing -verify-plugin github-branch-source

If any plugins matching the -verify-plugin string are found but don't
have valid wiki pages (currently that's the only thing we check
against, but if we add further criteria in the future, we'll check
those too), the script will notify the user and exit out with an exit
code of 1.