Ignore SCM API 2.0 related changes until the cause of JENKINS-41121
The basic idea of this is that the upgrades are sufficiently dangerous and *one
direction* (i.e. no rollback except for restoring from a pre-upgrade backup of
JENKINS_HOME) that it's best for Jenkins users to hide these versions until this
is figured out and made more safe.

See https://jenkins.io/blog/2017/01/17/scm-api-2/ for the list of affected
plugins.