Ban build pipeline plugin 2.0.0 - bundles JTH inside the plugin
https://issues.jenkins.io/browse/JENKINS-72353 reports that build pipeline
plugin 2.0.0 generates stack traces into the Jenkins log because it
incorrectly bundles the Jenkins test harness as a compile time dependency.

https://github.com/jenkinsci/build-pipeline-plugin/pull/131 fixes the
dependency.

I'm not entirely sure that this is a severe enough error to ban the
plugin version from distribution, especially since I am hopeful that
@dalvizu will deliver a 2.0.1 release very quickly.

I'm proposing this ban in case there is a delay in delivering the 2.0.1
release or in case the repeated stack traces are severe enough to justify
banning the plugin.