Mark SECURITY-3280 & SECURITY-2924 as fixed in ivy 2.6
https://www.jenkins.io/security/advisory/2023-09-06/#SECURITY-2924 is
fixed by upgrading the bundled version of ivy from 2.5.1 to 2.5.2.

https://github.com/jenkinsci/ivy-plugin/pull/87 is the pull request
that fixes SECURITY-2924.

https://www.jenkins.io/security/advisory/2023-09-06/#SECURITY-3093 is
fixed by adding the `@POST` annotation to the delete action.

https://github.com/jenkinsci/ivy-plugin/pull/98 is the pull request
that fixes SECURITY-3093.

I've run interactive tests with the 2.6 release of the Ivy plugin
using examples from the https://github.com/apache/ant-ivy repository
and the jobs are well behaved as far as I understand Apache Ivy.