Update to bouncycastle jdk18on with version 1.78.1
The Jenkins update-center2 was still using a bouncycastle version from
September 2008. Even though I exactly followed the instructions from the
update-center2's README.adoc file, I was not able to generate an OpenSSL
key (and certificate) that would work with that bouncycastle version.
Debugging into the library revealed that the old bouncycastle version
doesn't support reading private keys starting with the line
`-----BEGIN PRIVATE KEY-----`.

Instead, bouncycastle only supported one of the following starting lines
for private keys:

- `-----BEGIN RSA PRIVATE KEY-----`
- `-----BEGIN DSA PRIVATE KEY-----`
- `-----BEGIN EC PRIVATE KEY-----`

In the case of a starting line of `BEGIN PRIVATE KEY`, the
`readObject()` method of the `PEMReader` would just return `null` and
cause the update-center2 to crash.

However, the documented command `openssl genrsa -out demo.key 4096` also
generates a key file starting with `BEGIN PRIVATE KEY` when using modern
versions of OpenSSL (version 3.0.13 in my case).

For this reason, this change updates the used bouncycastle library to
the most recent version, with which I was able to sign the files again.