Suspend windows-slaves and multi-slave-config plugins
We suspend distribution of plugins that depend on a service that no
longer exists or is end of life like:

* https://github.com/jenkins-infra/update-center2/pull/732
* https://github.com/jenkins-infra/update-center2/pull/708
* https://github.com/jenkins-infra/update-center2/pull/676
* https://github.com/jenkins-infra/update-center2/pull/587

This is a similar case to services that are end of life.  Windows is
not end of life, but the services required for this plugin have been end
of life since at least March 2023. The plugin was deprecated as noted in
https://github.com/jenkinsci/windows-slaves-plugin?tab=readme-ov-file#notice-of-deprecation
almost two years ago in November 2022.
https://github.com/jenkinsci/windows-slaves-plugin/commit/cdafd13807267b535b13bbb25d9cc72376aa4594

The plugin installation statistics
https://stats.jenkins.io/pluginversions/windows-slaves.html show that
there are still almost 80,000 installations of the plugin, but only 22%
of those installations are on 2.426.3 or newer.  Most installations of
the plugin are not upgrading their Jenkins controller version.

Users with the plugin already installed will not be affected by
distribution being suspended.  Users with configuration as code
definitions that include the plugin will need to remove it from their
configuration as code definition.  They will receive an error message
when they attempt to download the plugin version.

Implied dependencies on the plugin were removed from Jenkins 2.386

* https://github.com/jenkinsci/jenkins/pull/7568

The plugin was removed from the setup wizard in Jenkins 2.379

* https://github.com/jenkinsci/jenkins/pull/7414

Fix https://github.com/jenkins-infra/helpdesk/issues/4221