Suspend 3 plugins that include test harness in the hpi file (#864)
Jenkins 2.434 and later will not load a plugin the includes Jenkins test harness because of the ways that Jenkins test harness harms a production Jenkins controller. Issues that highlight the problem include:

* https://issues.jenkins.io/browse/JENKINS-65650
* https://issues.jenkins.io/browse/JENKINS-66060
* https://issues.jenkins.io/browse/JENKINS-72353

More details are included in the Jenkins core pull request:

* https://github.com/jenkinsci/jenkins/pull/8714

Pull requests were suibmitted over a year ago for the three plugins.  There has been no response on any of those pull requests.

* https://github.com/jenkinsci/bmc-change-manager-imstm-plugin/pull/13
* https://github.com/jenkinsci/container-image-link-plugin/pull/110
* https://github.com/jenkinsci/itms-for-jira-plugin/pull/5

Installation counts as of Oct 2024:

* bmc-change-manager-imstm - 0
* container-image-link - 16
* itms-for-jira - 6

Testing done:

Confirmed that each of the plugins fail to load in Jenkins 2.492.3.