Remove deprecation of bmc-change-manager-imstm (#878)
The plugin was deprecated because it included the Jenkins test harness
in its most recent release.  The deprecation is in pull request:

* https://github.com/jenkins-infra/update-center2/pull/864

A maintainter has arrived, @mcohen11 , and has merged the pull request
that removes the Jenkins test harness.  That is pull request:

* https://github.com/jenkinsci/bmc-change-manager-imstm-plugin/pull/13

The plugin has been updated to require Java 17 and a modern Jenkins
version with pull request:

* https://github.com/jenkinsci/bmc-change-manager-imstm-plugin/pull/23

The new release is available from the update center as shown by:

https://get.jenkins.io/plugins/bmc-change-manager-imstm/66.v12c841920f7d/bmc-change-manager-imstm.hpi?mirrorlist

I've loaded the new release into Jenkins 2.519 successfully.

We can remove the deprecation.

This partially reverts 8280594e9e171d642abe6cf10db247436df62e7c